Skip the reserved-but-unformed sink slots in the criteria scan
sink_is_inhibited() only catches previously-real sinks that were later
removed (time_bin == time_bin_inhibited). The reserved capacity SWIFT
pre-allocates per top-level cell (Scheduler:cell_extra_sinks) uses a
different sentinel, time_bin_not_created, which the loop was not
checking. Once any real sink exists anywhere, the zero-sink guard added
in 928e1ae1b no longer applies, and every one of the reserved slots
(millions at production scale) was getting the full
sink_prepare_part_sink_formation_sink_criteria() call for every eligible
gas particle, instead of a cheap field check. Mirrors the same filter
already used in space.c's space_synchronize_sink_positions_mapper().

Also fix sink_prepare_part_sink_formation_sink_criteria() truncating the
sink age to int: sink_get_sink_age() returns internal time units, where
1 unit is of order a Gyr in SWIFT's standard unit system, so any sink
under about a Gyr old truncated to age 0 and never tripped the
age_threshold_unlimited "dead sink" short-circuit.

Both found during an independent red-team/physics review of the earlier
sink-formation fixes on this branch.

+++ src/sink/GEAR/sink.h
@@ -1360,7 +1360,7 @@ INLINE static void sink_prepare_part_sink_formation_sink_criteria(
 
   /* Determine if the sink is dead, i.e. if its age is bigger than the
      age_threshold_unlimited */
-  const int sink_age = sink_get_sink_age(sj, with_cosmology, cosmo, time);
+  const double sink_age = sink_get_sink_age(sj, with_cosmology, cosmo, time);
   char is_dead = sink_age > sink_props->age_threshold_unlimited;
 
   /* If the sink is dead, do not check the criteria for the si - p pair. */
